pyproject.toml
[project]
name = "protein-design-env"
version = "0.1.0"
description = "Simple protein design environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=1.0.0",
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "colorlog>=6.7.0",
    "tensorboard>=2.13.0",
    "hydra-core>=1.3.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.ruff]
target-version = "py310"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E", # PEP8 Errors
    "F", # Pyflakes (Unused imports, undefined names, etc.)
    "W", # Pycodestyle warnings
    "C90", # cyclomatic complexity, code complexity warnings
    "B", # Bugbear (for finding likely bugs and design issues)
    "Q", # Quilty (for checking quality issues in code)
    "S", # Security issues (e.g., Bandit checks)
    "I", # Import ordering (consistent import orders)
    "PT", # Pytest specific checks (if you are using pytest)
    "D", # Docstring conventions (ensures proper docstring formatting)
    "N", # Naming conventions (ensures consistent naming conventions)
]
ignore = [
    "E501", # Line too long: enforced by ruff-format, don't check again
    "E203", # Whitespace before ':' (black compatibility)
    "D104", # Missing docstring in public package
    "D100", # Missing docstring in public package
    "D107", # Missing docstring in `__init__`
    "D211", # No blank lines allowed before class docstring
    "PT011", # Ignore pytest exception too broad
    "S101", # Ignore Bandit checks for use of assert
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

# pytest configuration
[tool.pytest.ini_options]
minversion = "8.1"
addopts = "" # pytest-cov makes pycharm debugger unusable locally.
testpaths = ["tests"] # To speed up test collection.
python_files = "test_*.py" # To speed up test collection.
pythonpath = [".",
    "src",
    "scripts"
]

# Mypy configuration
[tool.mypy]
python_version = "3.10"
allow_untyped_decorators = true
disallow_subclassing_any = false
disallow_untyped_defs = true
no_implicit_optional = false
explicit_package_bases = true
follow_imports = "skip"
ignore_missing_imports = true
show_error_codes = true
strict = true
warn_return_any = true
warn_unused_ignores = true
modules = ["src", "tests"]
disable_error_code = ["call-overload"]


[dependency-groups]
dev = [
    "pre-commit==3.7.1",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "pytest-xdist==3.6.1",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"] # Ignore docstring for tests.
